新增查找所有房源

+++ src/main/java/com/chuangshu/livable/entity/IdInformation.java
@@ -33,7 +33,7 @@ public class IdInformation {
      * 证件号
      */
     @Column(name = "id_number")
-    private Integer idNumber;
+    private Long idNumber;
 
     /**
      * 证件照正面
